numa: numactl: this should take care of numactl -m 6-7 -N 6-7 $COMMAND
for users use a role template so that numactl runs generic exec (home) files with a auto type transition back to calling domain.

allow systemd to run numactl as-is. This should allow one to use numactl in service units (although the systemd domain my need some additional rules to support this)